refactor(gda): move script, resource, shader, theme groups into gda.commands (ADR-0040)
script.py absorbs script_run.py and the SCRIPT_RUN_COMMAND recipe;
ResourceReference stays in the shared core (multi-group shape, like
NodeProperty). Registration order and the public surface unchanged;
test imports follow the moves.
