+++ src/gda/errors.py
@@ -66,8 +66,6 @@ from gda.models import (
     OperationErrorEnvelope,
     PerfMonitorResult,
     PerfMonitorsResult,
-    ScriptDiagnostic,
-    ScriptValidateResult,
 )
 from gda.parser import parse_result
 from gda.runner import LaunchFailure, RunResult
@@ -297,79 +295,6 @@ def classify_run(
             f"structured-output contract violated: {exc}",
             result.stderr,
         )
-
-
-# A `SCRIPT ERROR: <message>` line and the `GDScript::reload (...:<line>)` frame
-# that follows it carry, between them, the one advisory diagnostic the engine
-# emits for a failed validate (issue #118). The line number is the final `:`
-# part of the reload frame; there is NO column on the standard build.
-# `[ \t]` (not `\s`) bounds the message capture so it cannot span newlines — an
-# empty SCRIPT ERROR message must not swallow the following reload frame.
-# Backtrace frames (`[n] _initialize (...)`) are operations.gd's own lines, not
-# the validated script's, so they are deliberately not matched.
-_SCRIPT_ERROR_LINE = re.compile(
-    r"^[ \t]*SCRIPT ERROR:[ \t]*(?P<message>.*?)[ \t]*$", re.MULTILINE
-)
-_RELOAD_FRAME = re.compile(r"GDScript::reload \([^)]*:(?P<line>\d+)\)")
-
-
-def parse_validate_diagnostics(stderr: str) -> list[ScriptDiagnostic]:
-    """Parse advisory ``script validate`` diagnostics from the engine's stderr.
-
-    A pure function (no engine, no I/O): the line/message of a failed
-    ``GDScript.reload()`` are available only from stderr, not from any bound API
-    (ADR-0002's stderr is advisory only — it is never parsed for the stable
-    success/failure outcome or error codes, only surfaced here as best-effort
-    diagnostics). ``column`` is always null (the engine exposes none).
-
-    The validate op does exactly ONE ``reload()``, so the only legitimate
-    ``GDScript::reload`` frame in stderr is the validated script's. Each
-    ``SCRIPT ERROR: <message>`` line is paired with a reload frame found ONLY in
-    the window up to the next ``SCRIPT ERROR`` line, and a ``SCRIPT ERROR`` with
-    no reload frame in that window is dropped: bounding the search keeps a later
-    error's frame from being mis-attributed to an earlier message, and the
-    frame requirement excludes unrelated engine ``SCRIPT ERROR`` noise — e.g. an
-    autoload's own startup error under ``--project``, whose frame is its
-    ``_init``/``_ready``, not ``GDScript::reload``. Returns ``[]`` when nothing
-    matches.
-    """
-    matches = list(_SCRIPT_ERROR_LINE.finditer(stderr))
-    diagnostics: list[ScriptDiagnostic] = []
-    for index, match in enumerate(matches):
-        window_end = (
-            matches[index + 1].start() if index + 1 < len(matches) else len(stderr)
-        )
-        frame = _RELOAD_FRAME.search(stderr, match.end(), window_end)
-        if frame is None:
-            continue
-        diagnostics.append(
-            ScriptDiagnostic(
-                line=int(frame.group("line")),
-                column=None,
-                message=match.group("message"),
-            )
-        )
-    return diagnostics
-
-
-def classify_script_validate(
-    result: RunResult, binary: Path
-) -> ScriptValidateResult | Failure:
-    """Classify the raw ``script validate`` result (issue #118).
-
-    The per-command layer for ``script validate``: the shared decision tree comes
-    from ``classify_run`` (an op error — missing path, wrong extension,
-    unreadable file — is still a ``Failure``). For a SUCCESSFUL op reporting an
-    invalid script (``valid=false``), the line/message diagnostics are not in the
-    sentinel — they live only in the engine's stderr — so this layer parses them
-    in and attaches them to the result.
-    """
-    outcome = classify_run(result, binary, ScriptValidateResult)
-    if isinstance(outcome, Failure):
-        return outcome
-    if not outcome.valid:
-        outcome.diagnostics = parse_validate_diagnostics(result.stderr)
-    return outcome
 
 
 def classify_info(result: RunResult, binary: Path) -> EngineVersion | Failure:
