+++ src/gda/models.py
@@ -442,529 +442,6 @@ class NodeProperty(BaseModel):
     )
 
 
-class ScriptCreateParams(BaseModel):
-    """The operation params of ``gda script create`` (issue #110).
-
-    ``path`` is the target ``.gd`` script file, addressed by its ``res://`` or
-    filesystem path (script-file addressing — by file path, not by
-    ``class_name``). ``content`` supplies verbatim source; when omitted, the
-    operation writes a minimal built-in template extending ``extends_type``.
-    ``content`` and ``extends_type`` are mutually exclusive at the CLI: verbatim
-    content is not templated, so a base class would have nowhere to go.
-    """
-
-    path: NormalizedPath = Field(description="Target .gd script path to write.")
-    content: str | None = Field(
-        default=None,
-        description=(
-            "Verbatim script source to write. When omitted, a minimal built-in "
-            "template extending 'extends_type' is written instead. Mutually "
-            "exclusive with the template's base class."
-        ),
-    )
-    extends_type: str | None = Field(
-        default=None,
-        description=(
-            "Base class for the built-in template's 'extends' line (e.g. Node, "
-            "Node2D). Ignored when 'content' is supplied; defaults to 'Node' "
-            "when neither is given."
-        ),
-    )
-
-    @model_validator(mode="after")
-    def _content_xor_extends(self) -> "ScriptCreateParams":
-        # Verbatim content is not templated, so a base class has nowhere to go:
-        # the two are mutually exclusive. Enforced model-side (ADR-0015) so the
-        # --params-json path rejects the conflict too, not just argv.
-        if self.content is not None and self.extends_type is not None:
-            raise ValueError("'content' and 'extends_type' are mutually exclusive.")
-        return self
-
-
-class ScriptCreateResult(BaseModel):
-    """The result of ``gda script create``: what was written where (issue #110).
-
-    Echoes the saved ``path`` and the ``class_name``/``extends`` the written
-    source declares, so an agent can assert the effect without a second call.
-    ``created_dirs`` lists parent directories the operation created before
-    saving, from outermost to innermost. The ``class_name``/``extends`` are
-    parsed from the written source.
-    """
-
-    path: str
-    class_name: str | None = Field(
-        default=None,
-        description=(
-            "The class_name the written script declares, or null when it declares none."
-        ),
-    )
-    extends: str | None = Field(
-        default=None,
-        description=(
-            "The base class the written script extends, or null when it declares none."
-        ),
-    )
-    created_dirs: list[str] = Field(
-        description=(
-            "Parent directories created before saving, from outermost to innermost."
-        )
-    )
-
-
-class ScriptGetParams(BaseModel):
-    """The operation params of ``gda script get``: the script file to read (issue #110).
-
-    ``path`` addresses the ``.gd`` script by its ``res://`` or filesystem path.
-    The source is read as raw text — the script is never loaded or compiled, so
-    reading it can never run project code (issue #30).
-    """
-
-    path: NormalizedPath = Field(description="The .gd script file to read.")
-
-
-class ScriptGetResult(BaseModel):
-    """The result of ``gda script get``: a script's source and metadata (issue #110).
-
-    Echoes the ``path``, the full ``source`` read as raw text, and the
-    ``class_name``/``extends`` the source declares (parsed from the text).
-    Carrying the source verbatim makes a ``create`` verifiable end-to-end:
-    ``create`` then ``get`` returns the same source.
-    """
-
-    path: str
-    source: str
-    class_name: str | None = Field(
-        default=None,
-        description=(
-            "The class_name the script declares, or null when it declares none."
-        ),
-    )
-    extends: str | None = Field(
-        default=None,
-        description=(
-            "The base class the script extends, or null when it declares none."
-        ),
-    )
-
-
-class ScriptListParams(BaseModel):
-    """The operation params of ``gda script list`` — none (ADR-0004).
-
-    ``script list`` enumerates the ``.gd`` scripts in the resolved project's
-    ``res://`` tree; the project is process context (``--project``), not an
-    operation param (ADR-0006), so the ``input`` schema is trivially empty.
-    """
-
-
-class ListedScript(BaseModel):
-    """One enumerated script of ``gda script list``: its path and metadata.
-
-    ``path`` is the script's ``res://`` path — the address an agent feeds back
-    into other script commands. ``class_name``/``extends`` are parsed cheaply
-    from the script's raw source (no compilation, issue #30); both are null when
-    the script declares neither, so the entry still names a file the listing
-    found rather than dropping it.
-    """
-
-    path: str
-    class_name: str | None = Field(
-        default=None,
-        description="The class_name the script declares, or null when it declares none.",
-    )
-    extends: str | None = Field(
-        default=None,
-        description="The base class the script extends, or null when it declares none.",
-    )
-
-
-class ScriptListResult(BaseModel):
-    """The result of ``gda script list``: the project's enumerated ``.gd`` scripts.
-
-    An empty project is a valid, empty listing — ``scripts == []`` — not a
-    failure.
-    """
-
-    scripts: list[ListedScript]
-
-
-class ScriptDeleteParams(BaseModel):
-    """The operation params of ``gda script delete``: the ``.gd`` file to remove."""
-
-    path: NormalizedPath = Field(description="The .gd script file to delete.")
-
-
-class ScriptDeleteResult(BaseModel):
-    """The result of ``gda script delete``: what was removed (issue #117).
-
-    Echoes the deleted script's ``path`` and the ``class_name``/``extends`` its
-    source declared (parsed from the raw text before deletion), so the result
-    names the content removed, not just the file path.
-    """
-
-    path: str
-    class_name: str | None = Field(
-        default=None,
-        description="The class_name the deleted script declared, or null when it declared none.",
-    )
-    extends: str | None = Field(
-        default=None,
-        description="The base class the deleted script extended, or null when it declared none.",
-    )
-
-
-class ScriptSetMode(str, Enum):
-    """The edit mode of ``gda script set``, the single source of truth (issue #133).
-
-    The CLI resolves exactly one mode from the supplied flags (its mutual-exclusion
-    check) and stamps it here, so the operation dispatches on this explicit
-    discriminator instead of re-inferring the mode from which params are present —
-    the inference precedence can no longer drift from the CLI's exclusivity rule.
-
-    - ``SEARCH_REPLACE`` — ``search``/``replace``: every literal (not regex)
-      occurrence of ``search`` is replaced with ``replace``.
-    - ``LINE_RANGE`` — ``start_line`` (+ optional ``end_line``) with ``content``:
-      the given 1-based, inclusive line span is replaced with ``content``.
-    - ``FULL`` — ``content`` only: the whole file is overwritten.
-    """
-
-    SEARCH_REPLACE = "search_replace"
-    LINE_RANGE = "line_range"
-    FULL = "full"
-
-
-def resolve_set_mode(
-    search: str | None,
-    replace: str | None,
-    start_line: int | None,
-    end_line: int | None,
-    content: str | None,
-) -> ScriptSetMode:
-    """Resolve a script/shader-set edit mode from the supplied params (issue #133).
-
-    The single home of the edit-mode rule, shared by ``script set`` and ``shader
-    set`` and by BOTH input paths (ADR-0015): exactly one of the three
-    mutually-exclusive modes must be supplied. Raises ``ValueError`` on a
-    violation — the CLI wrapper translates it to a usage error (exit 2) for argv,
-    while the params models surface it as the structured ``invalid_params`` for
-    ``--params-json``.
-    """
-    has_search = search is not None or replace is not None
-    has_line_range = start_line is not None or end_line is not None
-
-    if has_search:
-        if search is None or replace is None:
-            raise ValueError("'search' and 'replace' must be used together.")
-        if content is not None or has_line_range:
-            raise ValueError(
-                "'search'/'replace' cannot be combined with 'content', "
-                "'start_line', or 'end_line'."
-            )
-        return ScriptSetMode.SEARCH_REPLACE
-
-    if has_line_range:
-        if content is None:
-            raise ValueError("'start_line'/'end_line' require 'content'.")
-        if start_line is None:
-            raise ValueError("'end_line' requires 'start_line'.")
-        return ScriptSetMode.LINE_RANGE
-
-    if content is None:
-        raise ValueError(
-            "a set command needs an edit: 'search'/'replace', 'start_line' "
-            "(+ 'content'), or 'content' (full overwrite)."
-        )
-    return ScriptSetMode.FULL
-
-
-class ScriptSetParams(BaseModel):
-    """The operation params of ``gda script set`` (issue #118).
-
-    Edits an existing ``.gd`` script on disk as RAW TEXT — it never compiles or
-    loads the script, so editing one can never run project code (the read trust
-    boundary of issue #30). ``path`` addresses the script by its ``res://`` or
-    filesystem path. The remaining params carry one of three mutually-exclusive
-    edit modes; the CLI resolves which one and stamps it on ``mode`` (issue #133),
-    so the operation dispatches on that explicit discriminator rather than
-    re-inferring it from which params are present:
-
-    - **search-replace** (``mode = search_replace``) — ``search``/``replace`` both
-      present: every literal (not regex) occurrence of ``search`` is replaced with
-      ``replace``.
-    - **line-range** (``mode = line_range``) — ``start_line`` (+ optional
-      ``end_line``) with ``content``: the given 1-based, inclusive line span is
-      replaced with ``content``.
-    - **full** (``mode = full``) — only ``content`` present: the whole file is
-      overwritten.
-    """
-
-    path: NormalizedPath = Field(description="The .gd script file to edit.")
-    mode: ScriptSetMode | None = Field(
-        default=None,
-        description=(
-            "The resolved edit mode, the single source of truth the operation "
-            "dispatches on (issue #133). Derived model-side from the supplied "
-            "edit params (ADR-0015); a value passed in is ignored."
-        ),
-    )
-    search: str | None = Field(
-        default=None,
-        description=(
-            "search-replace mode: the literal substring to find (NOT a regex). "
-            "Every occurrence is replaced with 'replace'. Requires 'replace'."
-        ),
-    )
-    replace: str | None = Field(
-        default=None,
-        description=(
-            "search-replace mode: the literal text each occurrence of 'search' "
-            "is replaced with. Requires 'search'."
-        ),
-    )
-    start_line: int | None = Field(
-        default=None,
-        description=(
-            "line-range mode: the first line to replace, 1-based and inclusive. "
-            "Lines are the parts of the source split on '\\n', so a trailing "
-            "newline yields a final empty part: 'a\\nb\\n' is 3 lines "
-            "(['a', 'b', '']). Valid range is 1..N where N is that part count. "
-            "Requires 'content'."
-        ),
-    )
-    end_line: int | None = Field(
-        default=None,
-        description=(
-            "line-range mode: the last line to replace, 1-based and inclusive; "
-            "defaults to 'start_line' (a single-line replace). Must satisfy "
-            "start_line <= end_line <= N (the line count). Requires 'content'."
-        ),
-    )
-    content: str | None = Field(
-        default=None,
-        description=(
-            "The replacement text. In line-range mode it replaces the "
-            "start_line..end_line span; with no 'start_line' it overwrites the "
-            "entire file (full mode)."
-        ),
-    )
-
-    @model_validator(mode="after")
-    def _resolve_mode(self) -> "ScriptSetParams":
-        # Derive the edit mode from the supplied params (ADR-0015), so the argv
-        # and --params-json paths agree and a JSON caller cannot pass a mode
-        # inconsistent with the other edit fields.
-        self.mode = resolve_set_mode(
-            self.search, self.replace, self.start_line, self.end_line, self.content
-        )
-        return self
-
-
-class ScriptSetResult(BaseModel):
-    """The result of ``gda script set``: the edited script's metadata (issue #118).
-
-    Echoes the saved ``path`` and the ``class_name``/``extends`` re-parsed from
-    the source as written, so an edit round-trips through ``script get`` (the
-    verifier) without a second call — and an agent can assert the post-edit
-    metadata directly.
-    """
-
-    path: str
-    class_name: str | None = Field(
-        default=None,
-        description=(
-            "The class_name the edited source declares, or null when it declares none."
-        ),
-    )
-    extends: str | None = Field(
-        default=None,
-        description=(
-            "The base class the edited source extends, or null when it declares none."
-        ),
-    )
-
-
-class ScriptAttachParams(BaseModel):
-    """The operation params of ``gda script attach`` (issue #118).
-
-    Binds a ``.gd`` script to a node inside a ``.tscn`` scene: load the scene,
-    resolve the node by node path, attach the script, then re-pack and save. As a
-    scene mutation it instantiates the scene (the same inherent trust boundary as
-    ``node set``, ADR-0009): instantiating runs the ``_init`` of scripts already
-    attached in the scene, and for a script that compiles ``set_script``
-    constructs an instance of the newly-attached script, running its ``_init``
-    too. ``path`` is the scene; ``script`` is the ``.gd`` to attach. The script
-    must COMPILE: the headless engine silently rejects a non-compiling script
-    from ``set_script`` (it cannot be persisted into the scene), so attach
-    refuses one with ``script_compile_failed`` rather than report a phantom
-    success — check a script with ``script validate`` first.
-    """
-
-    path: NormalizedPath = Field(description="The .tscn scene file to mutate.")
-    node: str = Field(
-        description=(
-            "Node path relative to the scene root: '.' addresses the root "
-            "itself, 'Player/Arm' a nested node."
-        )
-    )
-    script: NormalizedPath = Field(
-        description="The .gd script file to attach to the node."
-    )
-
-
-class ScriptAttachResult(BaseModel):
-    """The result of ``gda script attach``: what was bound where (issue #118).
-
-    Echoes the ``scene_path``, the addressed ``node``, and the attached
-    ``script``, plus the script's ``class_name`` when it declares a global one —
-    the result an agent asserts to confirm the binding took effect, verifiable by
-    reading the saved scene back (the script now appears on the node).
-
-    ``attach`` is a mutation verb: it OVERWRITES an existing binding rather than
-    refusing it (issue #132). ``replaced_script`` makes that displacement visible
-    so the overwrite is never silent — an agent reads it to detect a clobber.
-    """
-
-    scene_path: str
-    node: str = Field(
-        description="The node path the script was attached to, relative to the scene root."
-    )
-    script: str = Field(description="The .gd script that was attached.")
-    class_name: str | None = Field(
-        default=None,
-        description=(
-            "The global class_name the attached script declares, or null when "
-            "it declares none."
-        ),
-    )
-    replaced_script: str | None = Field(
-        default=None,
-        description=(
-            "The resource_path of the script this attach DISPLACED, reported "
-            "verbatim — including a built-in/embedded script's sub-resource ref "
-            "(e.g. 'res://scene.tscn::GDScript_xxx'). Non-null whenever the node "
-            "already carried a script (attach overwrites-and-reports, issue "
-            "#132); null only when the node had no prior script."
-        ),
-    )
-
-
-class ScriptDiagnostic(BaseModel):
-    """One advisory diagnostic from ``gda script validate`` (issue #118).
-
-    Best-effort: parsed from the engine's stderr, not from a bound API, so it may
-    carry only the FIRST parse error. ``line`` is 1-based when the engine
-    reported it; ``column`` is ALWAYS null on the standard Godot build — the
-    engine does not expose a column for a parse error — and is kept as a field
-    only so the shape is stable if a future build ever does. ``message`` is the
-    engine's error text with its ``SCRIPT ERROR:`` prefix stripped.
-    """
-
-    line: int | None = Field(
-        default=None,
-        description="The 1-based source line the error was reported at, or null when unknown.",
-    )
-    column: int | None = Field(
-        default=None,
-        description=(
-            "Always null on the standard Godot build: the engine does not "
-            "expose a column for a parse error."
-        ),
-    )
-    message: str
-
-
-class ScriptValidateParams(BaseModel):
-    """The operation params of ``gda script validate``: the script to check (issue #118).
-
-    ``path`` addresses the ``.gd`` script by its ``res://`` or filesystem path.
-    Unlike the other script-file ops, validate DOES compile the script (it sets
-    the source on a fresh ``GDScript`` and reloads it to learn whether it parses),
-    but it never instantiates the script, so it does not run instance code. Pass
-    ``--project`` when the script extends a project ``class_name`` or preloads a
-    project resource and so needs project context to compile.
-    """
-
-    path: NormalizedPath = Field(description="The .gd script file to validate.")
-
-
-class ScriptValidateResult(BaseModel):
-    """The result of ``gda script validate``: whether the script compiles (issue #118).
-
-    Validating an INVALID script is a SUCCESSFUL operation — the command exits 0
-    and reports ``valid=false`` rather than failing. ``error_string`` carries the
-    engine's one-line summary of the compile error (null when valid).
-    ``diagnostics`` is a best-effort list parsed from the engine's stderr (the
-    only place line/message are available); it may hold only the first error, and
-    is empty when the script is valid or nothing could be parsed.
-    """
-
-    path: str
-    valid: bool = Field(
-        description="True when the script compiles (GDScript.reload() == OK), false otherwise."
-    )
-    error_string: str | None = Field(
-        default=None,
-        description=(
-            "The engine's one-line summary of the compile failure, or null when "
-            "the script is valid."
-        ),
-    )
-    diagnostics: list[ScriptDiagnostic] = Field(
-        default_factory=list,
-        description=(
-            "Best-effort advisory diagnostics parsed from the engine's stderr "
-            "(line + message). May hold only the first error; empty when valid."
-        ),
-    )
-
-
-class ResourceCreateParams(BaseModel):
-    """The operation params of ``gda resource create`` (issue #112).
-
-    ``path`` is the target ``.tres`` resource file, addressed by its ``res://``
-    or filesystem path (resource-file addressing — by file path). ``type`` is
-    the Resource type to instantiate and save: a built-in Resource class (e.g.
-    ``Gradient``, ``Curve``) OR a project-defined ``class_name`` (a GDScript
-    ``class_name Foo extends Resource``), resolved the same way ``node add``
-    resolves ``--type`` (issue #342) — mirroring ``scene create``'s ``root_type``
-    check against ``Node``.
-    """
-
-    path: NormalizedPath = Field(description="Target .tres resource path to write.")
-    type: str = Field(
-        description=(
-            "The Resource type to create: a built-in Resource class (e.g. "
-            "Gradient, Curve) or a registered Resource class_name (a GDScript "
-            "class_name Foo extends Resource)."
-        )
-    )
-
-
-class ResourceUidParams(BaseModel):
-    """The operation params of ``gda resource uid`` (issue #113).
-
-    Resolves a Godot resource UID to/from its resource path in BOTH directions
-    against the engine's UID cache — read-only, it never mutates the cache or any
-    file. ``target`` is the single addressing argument and selects the direction
-    by its form:
-
-    - a ``uid://…`` value: report the ``res://…`` path it resolves to.
-    - a ``res://…`` (or filesystem) path: report its assigned ``uid://…``.
-
-    The UID cache is the engine's own ``res://.godot/uid_cache.bin``, loaded at
-    startup, so resolution needs project context (``--project``); a projectless
-    run has no cache to query. This is distinct from ``.tres`` file CRUD: it
-    queries the cache, not a file's contents.
-    """
-
-    target: NormalizedPath = Field(
-        description=(
-            "The resolution target: a 'uid://…' value to resolve to its res:// "
-            "path, or a 'res://…' / filesystem path to resolve to its 'uid://…'. "
-            "The direction is chosen by whether 'target' begins with 'uid://'."
-        )
-    )
-
-
 # --- project static-analysis reads (issue #116) -----------------------------
 #
 # Four read-only, project-wide analysis commands, all backed by a single static
@@ -1178,432 +655,6 @@ class ExportRunResult(BaseModel):
     )
 
 
-class ScriptRunParams(BaseModel):
-    """The operation params of ``gda script run`` (issue #343, ADR-0031).
-
-    ``path`` is the ``res://`` path of the user script to run as a one-shot
-    ``godot --headless --path <project> --script <res://…>``. It is res://-only
-    (a res:// path resolves against the ``--project`` context, ADR-0006): an
-    absolute or non-``res://`` path is the ``invalid_path`` failure. A plain
-    ``str`` (NOT ``NormalizedPath``) because a res:// path is an engine-virtual
-    address, not a filesystem path — filesystem normalization would collapse the
-    ``res://`` double slash. The project is process context (``--project``), not
-    an operation param.
-    """
-
-    path: str = Field(
-        description="The res:// path of the script to run (e.g. res://tests/logic.gd)."
-    )
-
-
-class ScriptRunResult(BaseModel):
-    """The result of ``gda script run``: the user script's own run, passed through (ADR-0031).
-
-    This is the **public promotion of the internal Raw-run shape**
-    (:class:`gda.runner.RunResult`): a THIN boundary DTO built from a ``RunResult``
-    by dropping its ``launch_failure`` axis (that becomes the Error envelope) and
-    renaming ``exit_code`` → ``exit_status``. Unlike every other command,
-    ``script run`` does not interpret the user script's semantics — a deliberate
-    ``quit(1)`` is meaningful data the agent reads, not a gda failure — so this is
-    the **one** command whose *success* result can carry a non-zero
-    ``exit_status``. Agents must read ``exit_status`` and must not assume
-    ``success == zero``.
-
-    NOTE: a second passthrough consumer should promote this to a shared
-    ``RawRunResult`` model. Do NOT build that shared abstraction now: there is only
-    one consumer today (``export run`` returns a different domain shape — the
-    produced artifact — and does not reuse the raw run).
-    """
-
-    exit_status: int = Field(
-        description=(
-            "The user script's own process exit code, passed through verbatim — "
-            "non-zero (e.g. a deliberate quit(1)) is still a SUCCESS result, not a "
-            "gda failure (ADR-0031)."
-        )
-    )
-    stdout: str = Field(description="The script's standard output, captured verbatim.")
-    stderr: str = Field(description="The script's standard error, captured verbatim.")
-
-
-class ResourceCreateResult(BaseModel):
-    """The result of ``gda resource create``: what was written where (issue #112).
-
-    Echoes the saved ``path`` and the ``type`` of the resource it created, so an
-    agent can assert the effect (path + type) without a second call.
-    ``created_dirs`` lists parent directories the operation created before
-    saving, from outermost to innermost (mirrors ``scene``/``script`` create).
-    """
-
-    path: str
-    type: str = Field(description="The Godot resource class that was created.")
-    created_dirs: list[str] = Field(
-        description=(
-            "Parent directories created before saving, from outermost to innermost."
-        )
-    )
-
-
-class ShaderCreateParams(BaseModel):
-    """The operation params of ``gda shader create`` (issue #115).
-
-    ``path`` is the target ``.gdshader`` file, addressed by its ``res://`` or
-    filesystem path. A ``.gdshader`` is plain shader source authored as RAW
-    TEXT — the create/get/set trio authors the file directly and never loads or
-    compiles the shader AT THE OPERATION LEVEL (the same file-level boundary the
-    script group's create/get/set honor, issue #30). This bounds the operation,
-    not the run: like every command, a ``shader`` op still goes through the
-    headless runner, so resolving ``--project`` still constructs the project's
-    autoloads at engine startup (ADR-0009). ``content`` supplies verbatim shader
-    source; when omitted, the operation writes a minimal ``shader_type`` template.
-    """
-
-    path: NormalizedPath = Field(description="Target .gdshader path to write.")
-    content: str | None = Field(
-        default=None,
-        description=(
-            "Verbatim shader source to write. When omitted, a minimal template "
-            "declaring 'shader_type canvas_item' is written instead. Mutually "
-            "exclusive with the template's shader type."
-        ),
-    )
-    shader_type: str | None = Field(
-        default=None,
-        description=(
-            "Shader type for the built-in template's 'shader_type' line (e.g. "
-            "canvas_item, spatial, particles). Mutually exclusive with 'content' "
-            "(supplying both is rejected); defaults to 'canvas_item' when neither "
-            "is given."
-        ),
-    )
-
-    @model_validator(mode="after")
-    def _content_xor_shader_type(self) -> "ShaderCreateParams":
-        # Same rule as script create: verbatim content is not templated, so a
-        # shader type has nowhere to go. Enforced model-side (ADR-0015) so the
-        # --params-json path rejects the conflict, not just argv.
-        if self.content is not None and self.shader_type is not None:
-            raise ValueError("'content' and 'shader_type' are mutually exclusive.")
-        return self
-
-
-class ShaderCreateResult(BaseModel):
-    """The result of ``gda shader create``: what was written where (issue #115).
-
-    Echoes the saved ``path`` and the ``shader_type`` the written source
-    declares, so an agent can assert the effect without a second call.
-    ``created_dirs`` lists parent directories the operation created before
-    saving, from outermost to innermost. The ``shader_type`` is parsed from the
-    written source.
-    """
-
-    path: str
-    shader_type: str | None = Field(
-        default=None,
-        description=(
-            "The shader_type the written shader declares, or null when it "
-            "declares none."
-        ),
-    )
-    created_dirs: list[str] = Field(
-        description=(
-            "Parent directories created before saving, from outermost to innermost."
-        )
-    )
-
-
-class ResourceGetParams(BaseModel):
-    """The operation params of ``gda resource get``: the ``.tres`` to read (issue #112).
-
-    ``path`` addresses the resource by its ``res://`` or filesystem path. Loading
-    a ``.tres`` instantiates the resource (the same trust boundary every load
-    carries, ADR-0009), but a plain resource file holds data, not a script that
-    runs on load.
-    """
-
-    path: NormalizedPath = Field(description="The .tres resource file to read.")
-
-
-class ShaderGetParams(BaseModel):
-    """The operation params of ``gda shader get``: the shader file to read (issue #115).
-
-    ``path`` addresses the ``.gdshader`` by its ``res://`` or filesystem path.
-    The source is read as raw text — the operation itself never loads or compiles
-    the shader (the read boundary of issue #30). That bounds the operation, not
-    the run: like every command it goes through the headless runner, so resolving
-    ``--project`` still constructs the project's autoloads at engine startup
-    (ADR-0009).
-    """
-
-    path: NormalizedPath = Field(description="The .gdshader file to read.")
-
-
-class ResourceGetResult(BaseModel):
-    """The result of ``gda resource get``: a resource's properties as typed JSON (issue #112).
-
-    Echoes the ``path``, the resource's ``type`` (its engine class), and its
-    storage properties — the ones that serialize into the ``.tres`` — each as a
-    typed :class:`NodeProperty` (the same projection ``node get`` reports), so a
-    ``resource create`` round-trips: ``create`` then ``get`` reports the
-    resource it wrote.
-    """
-
-    path: str
-    type: str = Field(description="The resource's engine class (e.g. Gradient).")
-    properties: list[NodeProperty]
-
-
-class ResourceSetParams(BaseModel):
-    """The operation params of ``gda resource set`` (issue #120).
-
-    ``path`` is the ``.tres`` resource file to mutate, addressed by its ``res://``
-    or filesystem path; ``property`` is the resource property to set; ``value`` is
-    the CLI string value, coerced to the property's declared Godot type by the
-    operation (the same coercion rules as ``node set`` / ``project set``, #55)
-    before the ``.tres`` is re-saved. ``set`` edits an EXISTING property — an
-    unknown property is a clean error, never a silent create — so the declared
-    type to coerce to is always known (read off the resource's property list).
-    Mirrors ``project set`` closely: load → coerce to the declared type → save →
-    round-trip via ``resource get``.
-    """
-
-    path: NormalizedPath = Field(description="The .tres resource file to mutate.")
-    property: str = Field(
-        description="The resource property to set (e.g. interpolation_mode)."
-    )
-    value: str = Field(
-        description=(
-            "The value to set, as a string. The operation coerces it to the "
-            "property's declared Godot type (see the command catalog's 'Property "
-            "value coercion'). For Dictionary/Array JSON values, JSON integer "
-            "literals stay int and JSON float literals stay float; typed "
-            "containers assign entries through their declared container type. An "
-            "uncoercible value is a clean error."
-        )
-    )
-
-
-class ResourceSetResult(BaseModel):
-    """The result of ``gda resource set``: the one property it set (issue #120).
-
-    Echoes the ``path``, the ``property`` set, the declared ``type`` the CLI value
-    was coerced to, and the coerced ``value`` as JSON — the same projection
-    ``resource get`` reports for a storage property, so a ``set`` round-trips
-    through a ``get`` without re-reading the ``.tres``.
-    """
-
-    path: str
-    property: str
-    type: str = Field(
-        description="The property's declared Godot type the value was coerced to."
-    )
-    value: Any = Field(
-        description=(
-            "The coerced value as JSON, as the resource now holds it. "
-            + _OBJECT_SET_ECHO_DESC
-        ),
-        json_schema_extra=_projected_value_schema_extra,
-    )
-
-
-class ResourceDeleteParams(BaseModel):
-    """The operation params of ``gda resource delete``: the ``.tres`` file to remove (issue #120)."""
-
-    path: NormalizedPath = Field(description="The .tres resource file to delete.")
-
-
-class ResourceDeleteResult(BaseModel):
-    """The result of ``gda resource delete``: what was removed (issue #120).
-
-    Echoes the deleted resource's ``path`` and its ``type`` (the engine class,
-    read from the resource before deletion), so the result names the content
-    removed, not just the file path — mirroring ``scene``/``script delete``.
-    """
-
-    path: str
-    type: str = Field(
-        description="The deleted resource's engine class (e.g. Gradient)."
-    )
-
-
-class ResourceUidResult(BaseModel):
-    """The result of ``gda resource uid``: the resolved UID↔path pair (issue #113).
-
-    Both directions converge on the same shape — the resolved ``uid`` and the
-    ``path`` it maps to — so an agent always gets both sides of the mapping
-    regardless of which it queried. ``queried`` echoes which direction was
-    resolved, so the result is self-describing: ``uid`` means the target was a
-    ``uid://`` resolved to its path, ``path`` means the target was a path
-    resolved to its UID.
-    """
-
-    queried: str = Field(
-        description=(
-            "Which direction was resolved: 'uid' when the target was a 'uid://' "
-            "value (resolved to its path), 'path' when the target was a path "
-            "(resolved to its UID)."
-        )
-    )
-    uid: str = Field(description="The resource's 'uid://…' value.")
-    path: str = Field(description="The resource's 'res://…' path the UID maps to.")
-
-
-class ShaderGetResult(BaseModel):
-    """The result of ``gda shader get``: a shader's source and metadata (issue #115).
-
-    Echoes the ``path``, the full ``source`` read as raw text, and the
-    ``shader_type`` the source declares (parsed from the text). Carrying the
-    source verbatim makes a ``create`` verifiable end-to-end: ``create`` then
-    ``get`` returns the same source.
-    """
-
-    path: str
-    source: str
-    shader_type: str | None = Field(
-        default=None,
-        description=(
-            "The shader_type the shader declares, or null when it declares none."
-        ),
-    )
-
-
-class ShaderSetParams(BaseModel):
-    """The operation params of ``gda shader set`` (issue #115).
-
-    Edits an existing ``.gdshader`` on disk as RAW TEXT — the operation itself
-    never compiles or loads the shader (the edit boundary of issue #30). That
-    bounds the operation, not the run: like every command it goes through the
-    headless runner, so resolving ``--project`` still constructs the project's
-    autoloads at engine startup (ADR-0009). ``path`` addresses the shader by its
-    ``res://`` or filesystem path. The remaining params carry the SAME three
-    mutually-exclusive edit modes as ``script set`` (issue #118) — the shader
-    group reuses that edit interface rather than re-deriving it. The CLI resolves
-    which mode and stamps it on ``mode`` (issue #133), so the operation dispatches
-    on that explicit discriminator rather than re-inferring it from which params
-    are present:
-
-    - **search-replace** (``mode = search_replace``) — ``search``/``replace`` both
-      present: every literal (not regex) occurrence of ``search`` is replaced with
-      ``replace``.
-    - **line-range** (``mode = line_range``) — ``start_line`` (+ optional
-      ``end_line``) with ``content``: the given 1-based, inclusive line span is
-      replaced with ``content``.
-    - **full** (``mode = full``) — only ``content`` present: the whole file is
-      overwritten.
-    """
-
-    path: NormalizedPath = Field(description="The .gdshader file to edit.")
-    mode: ScriptSetMode | None = Field(
-        default=None,
-        description=(
-            "The resolved edit mode, the single source of truth the operation "
-            "dispatches on (issue #133). Derived model-side from the supplied "
-            "edit params (ADR-0015); a value passed in is ignored. The same edit "
-            "modes as script set (issue #118), reused here."
-        ),
-    )
-    search: str | None = Field(
-        default=None,
-        description=(
-            "search-replace mode: the literal substring to find (NOT a regex). "
-            "Every occurrence is replaced with 'replace'. Requires 'replace'."
-        ),
-    )
-    replace: str | None = Field(
-        default=None,
-        description=(
-            "search-replace mode: the literal text each occurrence of 'search' "
-            "is replaced with. Requires 'search'."
-        ),
-    )
-    start_line: int | None = Field(
-        default=None,
-        description=(
-            "line-range mode: the first line to replace, 1-based and inclusive. "
-            "Lines are the parts of the source split on '\\n', so a trailing "
-            "newline yields a final empty part: 'a\\nb\\n' is 3 lines "
-            "(['a', 'b', '']). Valid range is 1..N where N is that part count. "
-            "Requires 'content'."
-        ),
-    )
-    end_line: int | None = Field(
-        default=None,
-        description=(
-            "line-range mode: the last line to replace, 1-based and inclusive; "
-            "defaults to 'start_line' (a single-line replace). Must satisfy "
-            "start_line <= end_line <= N (the line count). Requires 'content'."
-        ),
-    )
-    content: str | None = Field(
-        default=None,
-        description=(
-            "The replacement text. In line-range mode it replaces the "
-            "start_line..end_line span; with no 'start_line' it overwrites the "
-            "entire file (full mode)."
-        ),
-    )
-
-    @model_validator(mode="after")
-    def _resolve_mode(self) -> "ShaderSetParams":
-        # Derive the edit mode from the supplied params (ADR-0015), so the argv
-        # and --params-json paths agree and a JSON caller cannot pass a mode
-        # inconsistent with the other edit fields.
-        self.mode = resolve_set_mode(
-            self.search, self.replace, self.start_line, self.end_line, self.content
-        )
-        return self
-
-
-class ShaderSetResult(BaseModel):
-    """The result of ``gda shader set``: the edited shader's metadata (issue #115).
-
-    Echoes the saved ``path`` and the ``shader_type`` re-parsed from the source
-    as written, so an edit round-trips through ``shader get`` (the verifier)
-    without a second call — and an agent can assert the post-edit metadata
-    directly.
-    """
-
-    path: str
-    shader_type: str | None = Field(
-        default=None,
-        description=(
-            "The shader_type the edited source declares, or null when it declares none."
-        ),
-    )
-
-
-class ThemeCreateParams(BaseModel):
-    """The operation params of ``gda theme create`` (issue #115).
-
-    ``path`` is the target ``.tres`` file. Unlike the shader trio (plain
-    file authoring), a Theme is an ENGINE-BACKED resource: the operation
-    constructs a ``Theme`` and writes it through ``ResourceSaver`` so the result
-    is a genuine, loadable ``.tres`` (verified by loading it back), not hand-
-    written text. The split mirrors the script group: file-level ops author text;
-    a resource-producing op goes through the engine.
-    """
-
-    path: NormalizedPath = Field(description="Target .tres Theme path to write.")
-
-
-class ThemeCreateResult(BaseModel):
-    """The result of ``gda theme create``: the created Theme resource (issue #115).
-
-    Echoes the saved ``path`` and the resource ``type`` written (``Theme``), so
-    an agent can assert the effect without a second call. ``created_dirs`` lists
-    parent directories the operation created before saving, from outermost to
-    innermost.
-    """
-
-    path: str
-    type: str = Field(description="The resource type written to the .tres (Theme).")
-    created_dirs: list[str] = Field(
-        description=(
-            "Parent directories created before saving, from outermost to innermost."
-        )
-    )
-
-
 class ResourceReference(BaseModel):
     """One referencing site found by ``gda project find-references`` (issue #116).
 
@@ -1617,6 +668,11 @@ class ResourceReference(BaseModel):
     project-level reference in ``project.godot``). ``context`` is the matched
     line/snippet, best-effort, so an agent can locate the reference without
     re-reading the file.
+
+    Stays in the shared core (ADR-0040 §4): its name says ``Resource`` but its
+    group does not — it is ``project find-references``'s result shape, so it
+    belongs beside :class:`ProjectFindReferencesResult`, not in the ``resource``
+    group's module.
     """
 
     path: str = Field(
