+++ tests/test_asset_file_commands.py
@@ -13,7 +13,7 @@ import json
 from typer.testing import CliRunner
 
 from gda.cli import app
-from gda.models import ScriptSetMode
+from gda.commands.script import ScriptSetMode
 from gda.runner import RunResult
 from tests.support import (
     SHADER_CREATE_RESULT,
