+++ tests/test_params_json.py
@@ -15,7 +15,7 @@ import typer
 from typer.testing import CliRunner
 
 from gda.cli import app
-from gda.models import ScriptSetMode
+from gda.commands.script import ScriptSetMode
 from gda.runner import RunResult
 from tests.support import (
     NODE_ADD_RESULT,
