+++ tests/test_render.py
@@ -24,6 +24,13 @@ from gda.commands.node import (
     render_node_set,
 )
 from gda.commands.scene import SceneNode
+from gda.commands.script import (
+    ListedScript,
+    ScriptCreateResult,
+    ScriptDeleteResult,
+    ScriptGetResult,
+    ScriptSetResult,
+)
 from gda.models import (
     DaemonStartResult,
     DaemonStatusResult,
@@ -32,20 +39,15 @@ from gda.models import (
     GameGetResult,
     GameRectResult,
     GameSetResult,
-    ListedScript,
     NodeProperty,
     PerfMonitor,
     PerfMonitorResult,
     PerfMonitorsResult,
     PerfPropertySample,
     PerfSignalEmission,
-    ScriptCreateResult,
-    ScriptDeleteResult,
-    ScriptGetResult,
-    ScriptSetResult,
 )
+from gda.commands.script import ScriptMetadata
 from gda.render import (
-    ScriptMetadata,
     format_value,
     render_daemon_start,
     render_daemon_status,
