+++ tests/test_resource_models.py
@@ -11,7 +11,7 @@ import json
 
 import jsonschema
 
-from gda.models import (
+from gda.commands.resource import (
     ResourceDeleteParams,
     ResourceDeleteResult,
     ResourceSetParams,
